% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{add.bias}
\alias{add.bias}
\title{Add bias column to data matrix}
\usage{
add.bias(x)
}
\arguments{
\item{x}{data matrix}
}
\description{
Add bias column to data matrix
}
\seealso{
Other stats: \code{\link{bstrap}}; \code{\link{center}};
  \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{rescale}}; \code{\link{sample.var}};
  \code{\link{sphere}}; \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

