% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{add_sql}
\alias{add_sql}
\title{Add additional SQL command or SET}
\usage{
add_sql(current_sql, new_sql)
}
\arguments{
\item{current_sql}{sql to be added to}

\item{new_sql}{sql to be added from}
}
\description{
Add additional SQL command or SET
}

