% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{as.numeric.cols}
\alias{as.numeric.cols}
\title{Cast specific columns as numeric}
\usage{
\method{as.numeric}{cols}(x, cols)
}
\arguments{
\item{x}{data.frame}

\item{cols}{column indicies to be converted to numeric}
}
\description{
Cast specific columns as numeric
}

