% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{audience_map}
\alias{audience_map}
\title{audience map
generated using ini files in sem-manager data}
\usage{
audience_map(sem_manager = "/data/sem-manager", tamgco = c("ta", "vi",
  "lf"))
}
\arguments{
\item{sem_manager}{director of sem-manager repo}

\item{tamgo}{which tamgcos to load}
}
\description{
audience map
generated using ini files in sem-manager data
}

