% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{avglast}
\alias{avglast}
\title{average last l values}
\usage{
avglast(x, l, order_by = NULL, ...)
}
\arguments{
\item{x}{data.frame}

\item{l}{integer number of values back to sum}

\item{order_by}{character}
}
\description{
average last l values
}

