% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/models.r
\name{bayes.exp.hier.mean}
\alias{bayes.exp.hier.mean}
\title{bayesian exponential estimate}
\usage{
bayes.exp.hier.mean(xbari, xbarij, ni, nij, alpha, beta)
}
\description{
bayesian exponential estimate
}

