% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{bstrap}
\alias{bstrap}
\title{bootstrap a statistic for a vector}
\usage{
bstrap(x, t = mean, n = 10000)
}
\arguments{
\item{x}{data vector}

\item{t}{statistic}

\item{n}{number of simulations}
}
\description{
bootstrap a statistic for a vector
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{center}};
  \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{rescale}}; \code{\link{sample.var}};
  \code{\link{sphere}}; \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

