% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cases.r
\name{cases}
\alias{cases}
\title{case statement
stolen from \textbf{memisc} package}
\usage{
cases(..., check.xor = c("warn", "stop", "ignore"))
}
\arguments{
\item{check.xor}{}
}
\description{
case statement
stolen from \textbf{memisc} package
}

