% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{center}
\alias{center}
\title{Center a numeric vector by subtracting off its mean.}
\usage{
center(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Center a numeric vector by subtracting off its mean.
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{ma}}; \code{\link{rescale}};
  \code{\link{sample.var}}; \code{\link{sphere}};
  \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

