% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{clean_sql}
\alias{clean_sql}
\title{clean query text}
\usage{
clean_sql(qry)
}
\arguments{
\item{qry}{query text loaded from read_sql}
}
\description{
clean query text
}
\seealso{
Other sql: \code{\link{add_sql}};
  \code{\link{collapse_char_vec}}; \code{\link{collap}};
  \code{\link{get_qry}}; \code{\link{get_tbl}};
  \code{\link{insert.sql}}; \code{\link{read_sql_raw}};
  \code{\link{read_sql}}; \code{\link{var_sub}}
}

