% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{coalesce}
\alias{coalesce}
\title{Coalesce two or more vectors
Replaces missing values of first vector in order of remaining vectors}
\usage{
coalesce(...)
}
\arguments{
\item{...}{vectors to coalesce}
}
\description{
Coalesce two or more vectors
Replaces missing values of first vector in order of remaining vectors
}
\seealso{
Other utils: \code{\link{any.na}};
  \code{\link{as.numeric.cols}}; \code{\link{chunk.ds}};
  \code{\link{format.percent}}; \code{\link{last}};
  \code{\link{relcumsum}}; \code{\link{relcum}};
  \code{\link{samp}}; \code{\link{send.email}};
  \code{\link{sort.abs}}; \code{\link{split.data}};
  \code{\link{which.na}}
}

