% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{collap}
\alias{collap}
\title{collapse vector for use with sql. Default output is string vector.}
\usage{
collap(x, type = c("character", "numeric", "custom"), custom = NULL, ...)
}
\arguments{
\item{x}{vector to be collapsed}

\item{type}{output of collapsed vector}

\item{custom}{optional. collapse using a custom string}

\item{...}{further arguments passed to paste}
}
\description{
collapse vector for use with sql. Default output is string vector.
}
\seealso{
Other sql: \code{\link{collapse_char_vec}};
  \code{\link{get_htbl}}; \code{\link{get_qry}};
  \code{\link{get_tbl}}; \code{\link{insert.sql}}
}

