% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{collapse_char_vec}
\alias{collapse_char_vec}
\title{alias 'to easily collapse (character) vector to hive (string) list'}
\usage{
collapse_char_vec(x)
}
\arguments{
\item{x}{vector to be collapsed}
}
\description{
alias 'to easily collapse (character) vector to hive (string) list'
}
\seealso{
Other sql: \code{\link{collap}}; \code{\link{get_htbl}};
  \code{\link{get_qry}}; \code{\link{get_tbl}};
  \code{\link{insert.sql}}
}

