% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.R
\name{collapse_vec}
\alias{collapse_vec}
\title{collapse_vec}
\usage{
collapse_vec(x, type = c("character", "numeric"), custom = NULL)
}
\description{
Collapse R vector for use with sql. Default output is string vector.
}

