% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{emperical.derivative}
\alias{emperical.derivative}
\title{Derivative estimation of 1d function
Uses change-in-y / chaing-in-x approximation}
\usage{
emperical.derivative(x, y)
}
\arguments{
\item{x}{input values}

\item{y}{output values}
}
\description{
Derivative estimation of 1d function
Uses change-in-y / chaing-in-x approximation
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{center}}; \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{rescale}}; \code{\link{sample.var}};
  \code{\link{sphere}}; \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

