% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{extrapolate_month}
\alias{extrapolate_month}
\title{expand monthly revenue to daily revenue}
\usage{
extrapolate_month(revenue_goal, month, year = "2016", day = "1")
}
\description{
expand monthly revenue to daily revenue
}

