% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{flast}
\alias{flast}
\title{apply function f to last l values}
\usage{
flast(x, f, l, order_by = NULL, ...)
}
\arguments{
\item{x}{data.frame}

\item{f}{function}

\item{l}{integer number of values back to sum}

\item{order_by}{character}
}
\description{
apply function f to last l values
}

