% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{fortnight}
\alias{fortnight}
\title{Calculate fortnight-in-year. Similar to lubridate::week}
\usage{
fortnight(x)
}
\arguments{
\item{x}{date-time}
}
\description{
Calculate fortnight-in-year. Similar to lubridate::week
}

