% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hive.r
\name{get_disk}
\alias{get_disk}
\title{get disk space
https://stat.ethz.ch/pipermail/r-help/2007-October/142319.html}
\usage{
get_disk(path = Sys.getenv("HOME"))
}
\description{
get disk space
https://stat.ethz.ch/pipermail/r-help/2007-October/142319.html
}
\seealso{
Other hive: \code{\link{get_hive_tbl}};
  \code{\link{get_hive}}; \code{\link{hive_system}};
  \code{\link{set_hive_cluster}};
  \code{\link{set_tempfile}};
  \code{\link{set_warehouse_path}}
}

