% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hive.r
\name{get_hive}
\alias{get_hive}
\title{retrieve hive query results as data.frame
WARNING: This function will throw an error if tabs appear in field data}
\usage{
get_hive(qry = NULL, init.qry = "", isfile = FALSE, queue = "sem")
}
\arguments{
\item{qry}{hive query to be executed (character vector)}

\item{init}{}
}
\description{
retrieve hive query results as data.frame
WARNING: This function will throw an error if tabs appear in field data
}
\seealso{
Other hive: \code{\link{get_hive_tbl}}
}

