% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hive.r
\name{get_hive_header}
\alias{get_hive_header}
\title{get column names of a hive table for tar.hive}
\usage{
get_hive_header(tbl)
}
\arguments{
\item{tbl}{data.frame}
}
\description{
get column names of a hive table for tar.hive
}

