% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{get_htbl}
\alias{get_htbl}
\title{get all rows from a table}
\usage{
get_htbl(x)
}
\arguments{
\item{x}{table name}
}
\description{
get all rows from a table
}
\seealso{
Other sql: \code{\link{collapse_char_vec}};
  \code{\link{collap}}; \code{\link{get_qry}};
  \code{\link{get_tbl}}; \code{\link{insert.sql}}
}

