% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{get_qry}
\alias{get_qry}
\title{get query}
\usage{
get_qry(con, qry, ...)
}
\arguments{
\item{con}{dbi connection}

\item{qry}{query text}
}
\description{
get query
}
\seealso{
Other sql: \code{\link{add_sql}}; \code{\link{clean_sql}};
  \code{\link{collapse_char_vec}}; \code{\link{collap}};
  \code{\link{get_htbl}}; \code{\link{get_tbl}};
  \code{\link{insert.sql}}; \code{\link{read_sql_raw}};
  \code{\link{read_sql}}; \code{\link{var_sub}}
}

