% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggacf}
\alias{ggacf}
\title{ggacf
\code{acf} plot using ggplot2
See: \url{http://stackoverflow.com/questions/17788859/acf-plot-with-ggplot2-setting-width-of-geom-bar}}
\usage{
ggacf(x, type = c("correlation", "covariance", "partial"), ci = 0.95, ...)
}
\description{
ggacf
\code{acf} plot using ggplot2
See: \url{http://stackoverflow.com/questions/17788859/acf-plot-with-ggplot2-setting-width-of-geom-bar}
}

