% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{gghist}
\alias{gghist}
\title{gghist
\code{hist} plot using ggplot2 using the \code{hist} default binwidth calculation}
\usage{
gghist(x)
}
\arguments{
\item{x}{Numeric data frame or vector to be plotted}
}
\value{
Histogram
}
\description{
gghist
\code{hist} plot using ggplot2 using the \code{hist} default binwidth calculation
}

