% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{hist.gg}
\alias{hist.gg}
\title{hist.gg}
\usage{
\method{hist}{gg}(x)
}
\arguments{
\item{x}{Numeric data frame or vector to be plotted}
}
\value{
Histogram
}
\description{
\code{hist} plot using ggplot2 using the \code{hist} default binwidth calculation
}

