% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hive.r
\name{hive_void}
\alias{hive_void}
\title{run hive command without return}
\usage{
hive_void(cmd, ...)
}
\arguments{
\item{cmd}{hive command}
}
\value{
void!
}
\description{
run hive command without return
}

