% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{importance.cv.glmnet}
\alias{importance.cv.glmnet}
\title{Variable importance for cv.glmnet fit
Importance determined by absolute value of mean coefficient}
\usage{
\method{importance}{cv.glmnet}(x, ...)
}
\arguments{
\item{x}{cv.glmnet fit}
}
\description{
Variable importance for cv.glmnet fit
Importance determined by absolute value of mean coefficient
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{center}}; \code{\link{emperical.derivative}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{rescale}}; \code{\link{sample.var}};
  \code{\link{sphere}}; \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

