% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{importance.glmnet}
\alias{importance.glmnet}
\title{Variable importance for glmnet fit
Importance determined by absolute value of mean coefficient}
\usage{
\method{importance}{glmnet}(x, ...)
}
\arguments{
\item{x}{glmnet fit}
}
\description{
Variable importance for glmnet fit
Importance determined by absolute value of mean coefficient
}

