% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{importance.lm}
\alias{importance.lm}
\title{Variable importance for lm fit
Importance determined by absolute value of t-statistic}
\usage{
\method{importance}{lm}(x, ...)
}
\arguments{
\item{x}{lm fit}
}
\description{
Variable importance for lm fit
Importance determined by absolute value of t-statistic
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{center}}; \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance}}; \code{\link{rescale}};
  \code{\link{sample.var}}; \code{\link{sphere}};
  \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

