% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{importance}
\alias{importance}
\title{Generic dispatch for importance}
\usage{
importance(x, ...)
}
\arguments{
\item{x}{model fit object with correct class}
}
\description{
Generic dispatch for importance
}

