% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.R
\name{insert.sql}
\alias{insert.sql}
\title{generate sql insert statement from dataframe}
\usage{
insert.sql(data, table.name)
}
\arguments{
\item{data}{dataframe to import}

\item{table.name}{name of table in db to insert into}
}
\description{
generate sql insert statement from dataframe
}

