% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kick_project}
\alias{kick_project}
\title{Create a Project with Standard Directories and Initialize Git}
\usage{
kick_project(
  name,
  root.dir = rstudioapi::readRStudioPreference("initial_working_directory", FALSE),
  gh = TRUE
)
}
\arguments{
\item{name}{The name of the main folder to create.}

\item{root.dir}{The root directory where the project will be created.}

\item{gh}{Logical indicating whether to set up a GitHub repository.}
}
\description{
Create a Project with Standard Directories and Initialize Git
}
