% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kick_tidy}
\alias{kick_tidy}
\title{Create a Tidy Project Directory Structure}
\usage{
kick_tidy(name)
}
\arguments{
\item{name}{The name of the main folder to create.}
}
\description{
Create a Tidy Project Directory Structure
}
