% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{lags}
\alias{lags}
\title{extend dplyr to multiple lags}
\usage{
lags(x, lags = 1L:2L, default = NA, order_by = NULL, ...)
}
\arguments{
\item{x}{a vector of values}

\item{lags}{vector of lags to be computed}

\item{default}{value used for non-existant rows. Defaults to \code{NA}.}

\item{order_by}{override the default ordering to use another vector}

\item{...}{Needed for compatibility with lag generic.}

\item{reduce}{function to reduce results}
}
\description{
extend dplyr to multiple lags
}

