% Generated by roxygen2 (4.0.2): do not edit by hand
\name{last}
\alias{last}
\title{Get last element of object.}
\usage{
last(x)
}
\description{
Get last element of object.
}
\seealso{
Other utils: \code{\link{any.na}}; \code{\link{percent}};
  \code{\link{samp}}; \code{\link{send_email}};
  \code{\link{which.na}}
}

