% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log.r
\name{log_error}
\alias{log_error}
\title{error logging}
\usage{
log_error(msg, quit = TRUE, ...)
}
\description{
error logging
}
\seealso{
Other log: \code{\link{log_info}};
  \code{\link{log_message}}; \code{\link{log_warn}};
  \code{\link{mem_change}}; \code{\link{mem_usage}};
  \code{\link{node_size}}; \code{\link{show_bytes}}
}

