% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log.r
\name{log_info}
\alias{log_info}
\title{info logging
alias for log_message}
\usage{
log_info(msg = "", ..., type = "INFO", showMem = TRUE)
}
\description{
info logging
alias for log_message
}
\seealso{
Other log: \code{\link{log_error}};
  \code{\link{log_message}}; \code{\link{log_warn}};
  \code{\link{mem_change}}; \code{\link{mem_usage}};
  \code{\link{node_size}}; \code{\link{show_bytes}}
}

