% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log.r
\name{log_message}
\alias{log_message}
\title{generic logging}
\usage{
log_message(msg = "", ..., type = "INFO", showMem = TRUE)
}
\arguments{
\item{msg}{}

\item{type}{}

\item{showMem}{}
}
\description{
generic logging
}
\seealso{
Other log: \code{\link{log_error}}; \code{\link{log_info}};
  \code{\link{log_warn}}; \code{\link{mem_change}};
  \code{\link{mem_usage}}; \code{\link{node_size}};
  \code{\link{show_bytes}}
}

