% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log.r
\name{log_warn}
\alias{log_warn}
\title{warn logging}
\usage{
log_warn(msg, ...)
}
\description{
warn logging
}
\seealso{
Other log: \code{\link{log_error}}; \code{\link{log_info}};
  \code{\link{log_message}}; \code{\link{mem_change}};
  \code{\link{mem_usage}}; \code{\link{node_size}};
  \code{\link{show_bytes}}
}

