% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{ma}
\alias{ma}
\title{moving average}
\usage{
ma(x, n, sides = 1)
}
\arguments{
\item{x}{a univariate or multivariate time series}

\item{n}{number of observations to include}

\item{sides}{if sides = 1 the filter coefficients are for past values only; if sides = 2 they are
centred around lag 0. In this case the length of the filter should be odd, but if it is even,
more of the filter is forward in time than backward.}
}
\description{
moving average
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{center}}; \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{rescale}}; \code{\link{sample.var}};
  \code{\link{sphere}}; \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

