% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log.r
\name{mem_change}
\alias{mem_change}
\title{Determine change in memory from running code}
\usage{
mem_change(code)
}
\arguments{
\item{code}{Code to evaluate.}
}
\value{
Change in memory (in megabytes) before and after running code.
}
\description{
Determine change in memory from running code
}
\examples{
# Need about 4 mb to store 1 million integers
mem_change(x <- 1:1e6)
# We get that memory back when we delete it
mem_change(rm(x))
}
\seealso{
Other log: \code{\link{log_error}}; \code{\link{log_info}};
  \code{\link{log_message}}; \code{\link{log_warn}};
  \code{\link{mem_usage}}; \code{\link{node_size}};
  \code{\link{show_bytes}}
}

