% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log.r
\name{mem_usage}
\alias{mem_usage}
\title{memory usage}
\usage{
mem_usage()
}
\value{
current r memory usage
}
\description{
memory usage
}
\seealso{
Other log: \code{\link{log_error}}; \code{\link{log_info}};
  \code{\link{log_message}}; \code{\link{log_warn}};
  \code{\link{mem_change}}; \code{\link{node_size}};
  \code{\link{show_bytes}}
}

