% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{model.binomial.bounds}
\alias{model.binomial.bounds}
\title{binomial model (proportion) lower/upper bounds for univariate rv}
\usage{
model.binomial.bounds(p = NULL, k = NULL, n = NULL, alpha = 0.05,
  type = c("normal", "wilson"))
}
\arguments{
\item{p}{observed or estimated proportion}

\item{k}{number of successes}

\item{n}{number of observations}

\item{alpha}{signifiance level}

\item{type}{method of approximation}
}
\value{
bounds
}
\description{
binomial model (proportion) lower/upper bounds for univariate rv
}

