% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/models.r
\name{model.normal.bounds}
\alias{model.normal.bounds}
\title{normal model (t test) using raw vector
lower/upper bounds for univariate rv with unknown var}
\usage{
model.normal.bounds(x, alpha = 0.05)
}
\arguments{
\item{x}{random variable}

\item{alpha}{significance level}
}
\value{
vector of lower/upper bounds of realizations of rv
}
\description{
normal model (t test) using raw vector
lower/upper bounds for univariate rv with unknown var
}
\seealso{
Other models: \code{\link{fit_bsm}};
  \code{\link{model.binomial.bounds}};
  \code{\link{model.normal.bounds.ss}}
}

