% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/models.r
\name{model.normal.bounds.ss}
\alias{model.normal.bounds.ss}
\title{normal model (t test) using summary statistics
lower/upper bounds for univariate rv with unknown var}
\usage{
model.normal.bounds.ss(xbar, sigma, n, alpha = 0.05)
}
\arguments{
\item{xbar}{sample mean}

\item{n}{number of observation}

\item{alpha}{significance level}

\item{sigma2}{sample variance}
}
\value{
vector of lower/upper bounds of realizations of rv
}
\description{
normal model (t test) using summary statistics
lower/upper bounds for univariate rv with unknown var
}
\seealso{
Other models: \code{\link{fit_bsm}};
  \code{\link{model.binomial.bounds}};
  \code{\link{model.normal.bounds}}
}

