% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log.r
\name{node_size}
\alias{node_size}
\title{node size}
\usage{
node_size()
}
\description{
node size
}
\seealso{
Other log: \code{\link{log_error}}; \code{\link{log_info}};
  \code{\link{log_message}}; \code{\link{log_warn}};
  \code{\link{mem_change}}; \code{\link{mem_usage}};
  \code{\link{show_bytes}}
}

