% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{nperiod}
\alias{nperiod}
\title{Calculate period-in-year where period can be defined by an arbitrary number of days (e.g. 14)}
\usage{
nperiod(x, ndays = 7)
}
\arguments{
\item{x}{date-time object}

\item{ndays}{number of days in period}
}
\description{
Calculate period-in-year where period can be defined by an arbitrary number of days (e.g. 14)
}

