% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_dkn}
\alias{parse_dkn}
\title{parse date_key_new (dkn)}
\usage{
parse_dkn(dkn, format = "\%Y\%m\%d")
}
\arguments{
\item{dkn}{int vector}

\item{format}{format to use}
}
\value{
Date object
}
\description{
parse date_key_new (dkn)
}

