% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hive.r
\name{parse_header}
\alias{parse_header}
\title{parse and shift first row as header}
\usage{
parse_header(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
parse and shift first row as header
}
\seealso{
Other hive: \code{\link{get_disk}};
  \code{\link{get_hive_tbl}}; \code{\link{get_hive}};
  \code{\link{hive_system}}; \code{\link{hive_void_file}};
  \code{\link{set_hive_cluster}};
  \code{\link{set_tempfile}};
  \code{\link{set_warehouse_path}}
}

