% Generated by roxygen2 (4.0.2): do not edit by hand
\name{percent}
\alias{percent}
\title{Convert proportion to percent string.}
\usage{
percent(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{Proportion to convert.}

\item{digits}{Number of digits to print.}

\item{format}{Output format.}
}
\description{
Convert proportion to percent string.
}
\seealso{
Other utils: \code{\link{any.na}}; \code{\link{last}};
  \code{\link{samp}}; \code{\link{send_email}};
  \code{\link{which.na}}
}

