% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/models.r
\name{ratio.var.est}
\alias{ratio.var.est}
\title{estimate Var[X/Y] using taylor series approximation
http://www.stat.cmu.edu/~hseltman/files/ratio.pdf}
\usage{
ratio.var.est()
}
\description{
estimate Var[X/Y] using taylor series approximation
http://www.stat.cmu.edu/~hseltman/files/ratio.pdf
}

