% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ini.r
\name{read_ini}
\alias{read_ini}
\title{read ini file
stolen from \textbf{raster} package}
\usage{
read_ini(filename, token = "=", commenttoken = ";", aslist = FALSE, case)
}
\arguments{
\item{filename}{Character. Filename of the .ini file}

\item{token}{Character. The character that separates the "name" (variable name) from the "value"}

\item{commenttoken}{Character. This token and everything that follows on the same line is considered a 'comment' that is not for machine consumption and is ignored in processing}

\item{aslist}{Logical. Should the values be returned as a list}

\item{case}{Optional. Function that operates on the text, such as toupper or tolower
This function allows for using inistrings that have "=" as part of a value (but the token cannot be part of the 'name' of a variable!). Sections can be missing.
Authors: Robert J. Hijmans
Date : October 2008
Version 0.9
Licence GPL v3}
}
\value{
A n*3 matrix of characters with columns: section, name, value; or a list if aslist=TRUE.
}
\description{
read ini file
stolen from \textbf{raster} package
}

