% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_query}
\alias{read_query}
\title{read_query
http://stackoverflow.com/questions/3580532/r-read-contents-of-text-file-as-a-query}
\usage{
read_query(file)
}
\description{
read_query
http://stackoverflow.com/questions/3580532/r-read-contents-of-text-file-as-a-query
}

