% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{read_sql_raw}
\alias{read_sql_raw}
\title{read raw sql query from file. Header is useful for setting hiveconf variables.
See: http://stackoverflow.com/questions/3580532/r-read-contents-of-text-file-as-a-query}
\usage{
read_sql_raw(file, header = NULL)
}
\arguments{
\item{file}{query file}

\item{header}{optional string to append to beginning of query, automatically adds ';'}
}
\description{
read raw sql query from file. Header is useful for setting hiveconf variables.
See: http://stackoverflow.com/questions/3580532/r-read-contents-of-text-file-as-a-query
}

