% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{relcum}
\alias{relcum}
\title{Relative cummulative summary}
\usage{
relcum(x)
}
\arguments{
\item{x}{cumsum vector}
}
\description{
Relative cummulative summary
}
\seealso{
Other utils: \code{\link{any.na}};
  \code{\link{as.numeric.cols}}; \code{\link{chunk.ds}};
  \code{\link{coalesce}}; \code{\link{format.percent}};
  \code{\link{last}}; \code{\link{relcumsum}};
  \code{\link{samp}}; \code{\link{send.email}};
  \code{\link{sort.abs}}; \code{\link{split.data}};
  \code{\link{which.na}}
}

