% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{repl.inf.data.frame}
\alias{repl.inf.data.frame}
\title{Replace Inf values in vector}
\usage{
repl.inf.data.frame(x, repl)
}
\arguments{
\item{x}{data.frame}
}
\description{
Replace Inf values in vector
}

