% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{repl.na}
\alias{repl.na}
\title{Replace NA values in data.frame}
\usage{
repl.na(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
Replace NA values in data.frame
}
\seealso{
Other utils: \code{\link{any.na}};
  \code{\link{as.numeric.cols}}; \code{\link{chunk.ds}};
  \code{\link{clear_all}}; \code{\link{coalesce}};
  \code{\link{compound_revenue}};
  \code{\link{extrapolate_days}};
  \code{\link{extrapolate_month}};
  \code{\link{format.percent}}; \code{\link{fortnight}};
  \code{\link{get_hostname}}; \code{\link{get_username}};
  \code{\link{lags}}; \code{\link{last}};
  \code{\link{nperiod}}; \code{\link{null_frame}};
  \code{\link{null_join}}; \code{\link{relcumsum}};
  \code{\link{relcum}}; \code{\link{safe_divide}};
  \code{\link{samp}}; \code{\link{send_email}};
  \code{\link{sort.abs}}; \code{\link{split.data}};
  \code{\link{sumlast}}; \code{\link{unload}};
  \code{\link{which.na}}
}

