% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{rescale}
\alias{rescale}
\title{Rescale a matrix or data frame
Standardise each column to have range [0, 1]}
\usage{
rescale(x)
}
\arguments{
\item{x}{data frame or matrix}
}
\value{
rescaled x
}
\description{
Rescale a matrix or data frame
Standardise each column to have range [0, 1]
}

