% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{sample.var}
\alias{sample.var}
\title{Sample variance (unbiased)}
\usage{
sample.var(x, x2)
}
\arguments{
\item{x}{data vector}

\item{x2}{data vector squared}
}
\description{
Sample variance (unbiased)
}

