% Generated by roxygen2 (4.0.2): do not edit by hand
\name{send_email}
\alias{send_email}
\title{#' Wrapper for sending email using mutt.}
\usage{
send_email(to, subject, body, attachment)
}
\description{
#' Wrapper for sending email using mutt.
}
\seealso{
Other utils: \code{\link{any.na}}; \code{\link{last}};
  \code{\link{percent}}; \code{\link{samp}};
  \code{\link{which.na}}
}

