% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hive.r
\name{set_hive_cluster}
\alias{set_hive_cluster}
\title{set hive cluster}
\usage{
set_hive_cluster(cluster = c("prod", "adhoc", "cdhtest", "core", "leo",
  "test"))
}
\arguments{
\item{cluster}{name of cluster to use with hive}
}
\value{
given cluster name
}
\description{
set hive cluster
}
\seealso{
Other hive: \code{\link{get_disk}};
  \code{\link{get_hive_tbl}}; \code{\link{get_hive}};
  \code{\link{hive_system}}; \code{\link{hive_void_file}};
  \code{\link{parse_header}}; \code{\link{set_tempfile}};
  \code{\link{set_warehouse_path}}
}

