% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hive.r
\name{set_warehouse_path}
\alias{set_warehouse_path}
\title{set warehouse path}
\usage{
set_warehouse_path(path)
}
\arguments{
\item{path}{}
}
\description{
set warehouse path
}
\seealso{
Other hive: \code{\link{get_disk}};
  \code{\link{get_hive_tbl}}; \code{\link{get_hive}};
  \code{\link{hive_system}};
  \code{\link{set_hive_cluster}};
  \code{\link{set_tempfile}}
}

