% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log.r
\name{show_bytes}
\alias{show_bytes}
\title{show bytes
byte size of r object}
\usage{
show_bytes(x)
}
\arguments{
\item{x}{object}
}
\value{
byte size as numeric
}
\description{
show bytes
byte size of r object
}
\seealso{
Other log: \code{\link{log_error}}; \code{\link{log_info}};
  \code{\link{log_message}}; \code{\link{log_warn}};
  \code{\link{mem_change}}; \code{\link{mem_usage}};
  \code{\link{node_size}}
}

