% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{sort.abs}
\alias{sort.abs}
\title{Sort by absolute value using base::sort}
\usage{
\method{sort}{abs}(x, ...)
}
\arguments{
\item{x}{vector be sorted}

\item{...}{arguments to be passed to or from methods or (for the
    default methods and objects without a class) to \code{sort.int}.}
}
\description{
Sort by absolute value using base::sort
}

