% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{sphere}
\alias{sphere}
\title{Sphere data. This is a stronger enforcement than scaling as it forces Cov = I.}
\usage{
sphere(x)
}
\arguments{
\item{x}{data matrix to be sphered}
}
\description{
Sphere data. This is a stronger enforcement than scaling as it forces Cov = I.
}

