% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{spline.derivative}
\alias{spline.derivative}
\title{Derivative estimation of 1d function}
\usage{
spline.derivative(x, y, deriv = 1, ...)
}
\arguments{
\item{x}{input values}

\item{y}{output values}
}
\description{
Derivative estimation of 1d function
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{center}}; \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{ma}}; \code{\link{rescale}};
  \code{\link{sample.var}}; \code{\link{sphere}};
  \code{\link{summarize.coefficient.matrix}}
}

