% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{split.data}
\alias{split.data}
\title{Split data into arbitrary partitions}
\usage{
\method{split}{data}(n, p)
}
\arguments{
\item{n}{number of indicies}

\item{p}{vector that sums to one}
}
\value{
List of split indicies
}
\description{
Split data into arbitrary partitions
}

