% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{summarize.coefficient.matrix}
\alias{summarize.coefficient.matrix}
\title{Summarize a coefficient matrix}
\usage{
summarize.coefficient.matrix(beta)
}
\arguments{
\item{beta}{matrix of coefficient where rows represent variables and columns represent hyperparameter}
}
\description{
Summarize a coefficient matrix
}

