% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{summarizeff}
\alias{summarizeff}
\title{dplyr summary simplifier
summarizes columns using a single aggregate function}
\usage{
summarizeff(x, f, metrics)
}
\description{
dplyr summary simplifier
summarizes columns using a single aggregate function
}

