% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_mdelhey}
\alias{theme_mdelhey}
\title{theme_mdelhey
Minimal ggplot2 theme based off of theme_tufte and theme_bw}
\usage{
theme_mdelhey(base_size = 12, base_family = "sans", ticks = TRUE)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{ticks}{}
}
\description{
theme_mdelhey
Minimal ggplot2 theme based off of theme_tufte and theme_bw
}

