% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_tufte}
\alias{theme_tufte}
\title{theme_tufte
\url{https://github.com/mattdelhey/ggthemes/blob/master/R/tufte.R}}
\usage{
theme_tufte(base_size = 11, base_family = "serif", ticks = TRUE)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{ticks}{}
}
\description{
theme_tufte
\url{https://github.com/mattdelhey/ggthemes/blob/master/R/tufte.R}
}

