% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{transfer}
\alias{transfer}
\title{transfer from swt to tamp}
\usage{
transfer(con, tn_in, tn_out, swtparams = "../swtparams.r")
}
\arguments{
\item{con}{output database connection}

\item{tn_in}{input table name}

\item{tn_out}{output table name}

\item{swtparam}{location of sem warehouse param file}
}
\description{
transfer from swt to tamp
}

