% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log.r
\name{try_log}
\alias{try_log}
\title{try logging
Try something and exit with a log message if it doesn't work}
\usage{
try_log(expr, err.msg = "")
}
\arguments{
\item{expr}{}

\item{err.msg}{}
}
\description{
try logging
Try something and exit with a log message if it doesn't work
}

