% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unload}
\alias{unload}
\title{unload
http://stackoverflow.com/questions/6979917/how-to-unload-a-package-without-restarting-r}
\usage{
unload(package)
}
\description{
unload
http://stackoverflow.com/questions/6979917/how-to-unload-a-package-without-restarting-r
}

