% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{var_sub}
\alias{var_sub}
\title{substitute variables in string}
\usage{
var_sub(string, ...)
}
\arguments{
\item{string}{input string with variable}

\item{...}{variables to replaced of the form "blah" = 5}
}
\description{
substitute variables in string
}
\seealso{
Other sql: \code{\link{add_sql}}; \code{\link{clean_sql}};
  \code{\link{collapse_char_vec}}; \code{\link{collap}};
  \code{\link{get_htbl}}; \code{\link{get_qry}};
  \code{\link{get_tbl}}; \code{\link{insert.sql}};
  \code{\link{read_sql_raw}}; \code{\link{read_sql}}
}

