% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{viator_audience_map}
\alias{viator_audience_map}
\title{viator audience map
generated manually from google adwords}
\usage{
viator_audience_map()
}
\description{
viator audience map
generated manually from google adwords
}

