% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{vsub}
\alias{vsub}
\title{Variable substitution
Substitute named variables}
\usage{
vsub(txt, clean = FALSE, ...)
}
\description{
Variable substitution
Substitute named variables
}
\examples{
vsub("test_\%s1_\%s2", "\%s1" = 6, "\%s2" = 10)
}

