% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{which.inf}
\alias{which.inf}
\title{Subset a vector to its values which are INF}
\usage{
which.inf(x)
}
\arguments{
\item{x}{Vector with missing values}
}
\value{
Subset of vector without missing values
}
\description{
Subset a vector to its values which are INF
}

