% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xl.r
\name{xl_create}
\alias{xl_create}
\title{create a workbook}
\usage{
xl_create(fname, overwrite = TRUE)
}
\arguments{
\item{fname}{'character' output file name}

\item{overwrite}{'logical' overwrite if file exists?}
}
\value{
XLConnect workbook
}
\description{
create a workbook
}

