% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xl.r
\name{xl_write}
\alias{xl_write}
\title{Write data to region}
\usage{
xl_write(wb, data, sheet, region, where = "$A$1", overwrite = TRUE,
  header = TRUE, rownames = NULL)
}
\arguments{
\item{wb}{XLConnect workbook}

\item{data}{data.frame}

\item{sheet}{'character' sheet to write in}

\item{region}{'character' excel region to name}

\item{where}{'character' excel location formula}

\item{overwrite}{'logical' overwrite region name?}

\item{header}{'logical' write header?}

\item{rownames}{Name (\code{character}) of column to use for the row names of the provided
  \code{data} object. If specified, the row names of the data object (\code{data.frame}) will be
  included as an additional column with the specified name. If \code{rownames = NULL} (default),
  no row names will be included. May also be a \code{list} in case multiple data objects are
  written in one call (see below).}
}
\description{
Write data to region
}

