% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{any.na}
\alias{any.na}
\title{For each col in a df, see if there are any NA's in that col.}
\usage{
any.na(x)
}
\arguments{
\item{df}{Dataframe to check.}

\item{dimn}{Dimension to apply on. See ?apply.}
}
\value{
Logical vector of length = number of columns.
}
\description{
For each col in a df, see if there are any NA's in that col.
}

