% Generated by roxygen2 (4.0.2): do not edit by hand
\name{any.na}
\alias{any.na}
\title{For each col in a df, see if there are any NA's in that col.}
\usage{
\method{any}{na}(df, dimn = NULL)
}
\arguments{
\item{df}{Dataframe to check.}

\item{dimn}{Dimension to apply on. See ?apply.}
}
\value{
Logical vector of length = number of columns.
}
\description{
For each col in a df, see if there are any NA's in that col.
}
\seealso{
Other utils: \code{\link{last}}; \code{\link{percent}};
  \code{\link{samp}}; \code{\link{send_email}};
  \code{\link{which.na}}
}

