% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{as.numeric.cols}
\alias{as.numeric.cols}
\title{Cast specific columns as numeric}
\usage{
\method{as.numeric}{cols}(x, cols)
}
\arguments{
\item{x}{data.frame}

\item{cols}{column indicies to be converted to numeric}
}
\description{
Cast specific columns as numeric
}
\seealso{
Other utils: \code{\link{any.na}}; \code{\link{chunk.ds}};
  \code{\link{coalesce}}; \code{\link{format.percent}};
  \code{\link{last}}; \code{\link{relcumsum}};
  \code{\link{relcum}}; \code{\link{samp}};
  \code{\link{send.email}}; \code{\link{sort.abs}};
  \code{\link{split.data}}; \code{\link{which.na}}
}

