% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{bstrap}
\alias{bstrap}
\title{bootstrap a statistic for a vector}
\usage{
bstrap(x, t = mean, n = 10000)
}
\arguments{
\item{x}{data vector}

\item{t}{statistic}

\item{n}{number of simulations}
}
\description{
bootstrap a statistic for a vector
}

