% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{chunk.ds}
\alias{chunk.ds}
\title{Chunk date range into given size. The last chunk may contain fewer elements.}
\usage{
chunk.ds(dt.range = NULL, dt.start = NULL, dt.end = NULL, chunk.size,
  reverse = FALSE)
}
\arguments{
\item{dt.range}{vector of dates}

\item{dt.start}{first day for vector construction}

\item{dt.end}{last day for vector construction}

\item{chunk.size}{size of each chunk}
}
\description{
Chunk date range into given size. The last chunk may contain fewer elements.
}
\seealso{
Other utils: \code{\link{any.na}};
  \code{\link{as.numeric.cols}}; \code{\link{clear_all}};
  \code{\link{coalesce}}; \code{\link{compound_revenue}};
  \code{\link{extrapolate_days}};
  \code{\link{extrapolate_month}};
  \code{\link{format.percent}}; \code{\link{fortnight}};
  \code{\link{get_hostname}}; \code{\link{get_username}};
  \code{\link{lags}}; \code{\link{last}};
  \code{\link{nperiod}}; \code{\link{null_frame}};
  \code{\link{null_join}}; \code{\link{rand}};
  \code{\link{relcumsum}}; \code{\link{relcum}};
  \code{\link{repl.inf.data.frame}}; \code{\link{repl.na}};
  \code{\link{safe_divide}}; \code{\link{samp}};
  \code{\link{send_email}}; \code{\link{sort.abs}};
  \code{\link{split.data}}; \code{\link{unload}};
  \code{\link{vsub}}; \code{\link{which.inf}};
  \code{\link{which.na}}
}

