% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{format.percent}
\alias{format.percent}
\title{Convert proportion to percent string.}
\usage{
\method{format}{percent}(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{Proportion to convert.}

\item{digits}{Number of digits to print.}

\item{format}{Output format.}
}
\description{
Convert proportion to percent string.
}

