% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{format.percent}
\alias{format.percent}
\title{Convert proportion to percent string.}
\usage{
\method{format}{percent}(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{Proportion to convert.}

\item{digits}{Number of digits to print.}

\item{format}{Output format.}
}
\description{
Convert proportion to percent string.
}
\seealso{
Other utils: \code{\link{any.na}}; \code{\link{chunk.ds}};
  \code{\link{last}}; \code{\link{samp}};
  \code{\link{send.email}}; \code{\link{split.data}};
  \code{\link{which.na}}
}

