% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hive.r
\name{get_hive_tbl}
\alias{get_hive_tbl}
\title{read hive table to r data.frame}
\usage{
get_hive_tbl(tbl, add.qry = "", init.qry = "")
}
\arguments{
\item{tbl}{hive table name}
}
\description{
read hive table to r data.frame
}
\seealso{
Other hive: \code{\link{get_disk}}; \code{\link{get_hive}};
  \code{\link{hive_system}}; \code{\link{hive_void_file}};
  \code{\link{parse_header}};
  \code{\link{set_hive_cluster}};
  \code{\link{set_tempfile}};
  \code{\link{set_warehouse_path}}
}

