% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{get_tbl}
\alias{get_tbl}
\title{get all rows from a table}
\usage{
get_tbl(con, x)
}
\arguments{
\item{con}{dbi connection}

\item{x}{table name}
}
\description{
get all rows from a table
}
\seealso{
Other sql: \code{\link{add_sql}}; \code{\link{clean_sql}};
  \code{\link{collapse_char_vec}}; \code{\link{collap}};
  \code{\link{insert.sql}}; \code{\link{read_sql_raw}};
  \code{\link{read_sql}}; \code{\link{var_sub}}
}

