% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggcolor}
\alias{ggcolor}
\title{Recreate ggplot2 colors for arbitrary number of factors (stackoverflow). http://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
\usage{
ggcolor(n)
}
\description{
Recreate ggplot2 colors for arbitrary number of factors (stackoverflow). http://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}

