% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggcolor}
\alias{ggcolor}
\title{ggcolor}
\usage{
ggcolor(n)
}
\description{
Recreate ggplot2 colors for arbitrary number of factors
See: \url{http://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
}

