% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hive.r
\name{hive}
\alias{hive}
\title{run hive command}
\usage{
hive(cmd, fileout = tempfile("Rtmp_", fileext = ".dat"), isfile = FALSE)
}
\arguments{
\item{cmd}{}

\item{fileout}{}

\item{isfile}{}
}
\value{
stdout
}
\description{
run hive command
}

