% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{importance.cv.glmnet}
\alias{importance.cv.glmnet}
\title{Variable importance for cv.glmnet fit
Importance determined by absolute value of mean coefficient}
\usage{
\method{importance}{cv.glmnet}(x, ...)
}
\arguments{
\item{x}{cv.glmnet fit}
}
\description{
Variable importance for cv.glmnet fit
Importance determined by absolute value of mean coefficient
}

