% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{importance.lm}
\alias{importance.lm}
\title{Variable importance for lm fit
Importance determined by absolute value of t-statistic}
\usage{
\method{importance}{lm}(x, ...)
}
\arguments{
\item{x}{lm fit}
}
\description{
Variable importance for lm fit
Importance determined by absolute value of t-statistic
}

