% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{importance}
\alias{importance}
\title{Generic dispatch for importance}
\usage{
importance(x, ...)
}
\arguments{
\item{x}{model fit object with correct class}
}
\description{
Generic dispatch for importance
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{center}}; \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{rescale}};
  \code{\link{sample.var}}; \code{\link{sphere}};
  \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

