% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{insert.sql}
\alias{insert.sql}
\title{generate sql insert statement from dataframe}
\usage{
insert.sql(data, table.name)
}
\arguments{
\item{data}{dataframe to import}

\item{table.name}{name of table in db to insert into}
}
\description{
generate sql insert statement from dataframe
}
\seealso{
Other sql: \code{\link{add_sql}}; \code{\link{clean_sql}};
  \code{\link{collapse_char_vec}}; \code{\link{collap}};
  \code{\link{get_htbl}}; \code{\link{get_qry}};
  \code{\link{get_tbl}}; \code{\link{read_sql_raw}};
  \code{\link{read_sql}}; \code{\link{var_sub}}
}

