% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{lags}
\alias{lags}
\title{extend dplyr to multiple lags}
\usage{
lags(x, lags = 1L:2L, default = NA, order_by = NULL, ...)
}
\arguments{
\item{x}{a vector of values}

\item{lags}{vector of lags to be computed}

\item{default}{value used for non-existant rows. Defaults to \code{NA}.}

\item{order_by}{override the default ordering to use another vector}

\item{...}{Needed for compatibility with lag generic.}

\item{reduce}{function to reduce results}
}
\description{
extend dplyr to multiple lags
}
\seealso{
Other utils: \code{\link{any.na}};
  \code{\link{as.numeric.cols}}; \code{\link{chunk.ds}};
  \code{\link{clear_all}}; \code{\link{coalesce}};
  \code{\link{compound_revenue}};
  \code{\link{extrapolate_days}};
  \code{\link{extrapolate_month}};
  \code{\link{format.percent}}; \code{\link{fortnight}};
  \code{\link{get_hostname}}; \code{\link{get_username}};
  \code{\link{last}}; \code{\link{nperiod}};
  \code{\link{null_frame}}; \code{\link{null_join}};
  \code{\link{relcumsum}}; \code{\link{relcum}};
  \code{\link{repl.na}}; \code{\link{safe_divide}};
  \code{\link{samp}}; \code{\link{send_email}};
  \code{\link{sort.abs}}; \code{\link{split.data}};
  \code{\link{sumlast}}; \code{\link{unload}};
  \code{\link{which.na}}
}

