% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{last}
\alias{last}
\title{Get last element of object.}
\usage{
last(x)
}
\description{
Get last element of object.
}
\seealso{
Other utils: \code{\link{any.na}}; \code{\link{chunk.ds}};
  \code{\link{format.percent}}; \code{\link{samp}};
  \code{\link{send.email}}; \code{\link{split.data}};
  \code{\link{which.na}}
}

