% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{null_frame}
\alias{null_frame}
\title{create a null data frame that contains a value for each combination of factors}
\usage{
null_frame(x, factors, measures, fill = NA)
}
\arguments{
\item{x}{data frame}

\item{factors}{vector of factor column names; filled by unique values in data frame}

\item{measures}{vector of measures column names; filled by "fill"}

\item{fill}{default fill value for measure columns}
}
\description{
create a null data frame that contains a value for each combination of factors
}
\seealso{
Other utils: \code{\link{any.na}};
  \code{\link{as.numeric.cols}}; \code{\link{chunk.ds}};
  \code{\link{clear_all}}; \code{\link{coalesce}};
  \code{\link{compound_revenue}};
  \code{\link{extrapolate_days}};
  \code{\link{extrapolate_month}};
  \code{\link{format.percent}}; \code{\link{fortnight}};
  \code{\link{get_hostname}}; \code{\link{get_username}};
  \code{\link{lags}}; \code{\link{last}};
  \code{\link{nperiod}}; \code{\link{null_join}};
  \code{\link{relcumsum}}; \code{\link{relcum}};
  \code{\link{repl.na}}; \code{\link{safe_divide}};
  \code{\link{samp}}; \code{\link{send_email}};
  \code{\link{sort.abs}}; \code{\link{split.data}};
  \code{\link{sumlast}}; \code{\link{unload}};
  \code{\link{which.na}}
}

