% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.R
\name{read_query}
\alias{read_query}
\title{read_query}
\usage{
read_query(file, header = NULL)
}
\description{
Read sql query from file. Header is useful for setting hiveconf variables.
See: http://stackoverflow.com/questions/3580532/r-read-contents-of-text-file-as-a-query
}

