% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{read_sql}
\alias{read_sql}
\title{read sql query from file WITH variables and cleaning}
\usage{
read_sql(file, header = NULL, ...)
}
\arguments{
\item{file}{query file}

\item{header}{optional string to append to beginning of query, automatically adds ';'}

\item{...}{variables to replaced of the form "blah" = 5}
}
\description{
read sql query from file WITH variables and cleaning
}
\seealso{
Other sql: \code{\link{add_sql}}; \code{\link{clean_sql}};
  \code{\link{collapse_char_vec}}; \code{\link{collap}};
  \code{\link{get_tbl}}; \code{\link{insert.sql}};
  \code{\link{read_sql_raw}}; \code{\link{var_sub}}
}

