% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{rescale}
\alias{rescale}
\title{Rescale a matrix or data frame
Standardise each column to have range [0, 1]}
\usage{
rescale(x)
}
\arguments{
\item{x}{data frame or matrix}
}
\value{
rescaled x
}
\description{
Rescale a matrix or data frame
Standardise each column to have range [0, 1]
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{center}}; \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{ma}}; \code{\link{sample.var}};
  \code{\link{sphere}}; \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

