% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{safe_divide}
\alias{safe_divide}
\title{safe divide}
\usage{
safe_divide(x, y, fail = 0)
}
\arguments{
\item{x}{numerator}

\item{y}{denominator}

\item{fail}{replacement for when denominator is zero}
}
\description{
safe divide
}

