% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{samp}
\alias{samp}
\title{Sample n rows from a dataframe. Works just like head/tail.}
\usage{
samp(x, n = 6L)
}
\arguments{
\item{x}{Dataframe to sample.}

\item{n}{Number of rows to sample.}
}
\value{
Dataframe with n randomly selected rows.
}
\description{
Sample n rows from a dataframe. Works just like head/tail.
}

