% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{samp}
\alias{samp}
\title{Sample n rows from a dataframe. Works just like head/tail.}
\usage{
samp(x, n = 6L)
}
\arguments{
\item{x}{Dataframe to sample.}

\item{n}{Number of rows to sample.}
}
\value{
Dataframe with n randomly selected rows.
}
\description{
Sample n rows from a dataframe. Works just like head/tail.
}
\seealso{
Other utils: \code{\link{any.na}};
  \code{\link{as.numeric.cols}}; \code{\link{chunk.ds}};
  \code{\link{clear_all}}; \code{\link{coalesce}};
  \code{\link{compound_revenue}};
  \code{\link{extrapolate_days}};
  \code{\link{extrapolate_month}};
  \code{\link{format.percent}}; \code{\link{fortnight}};
  \code{\link{get_hostname}}; \code{\link{get_username}};
  \code{\link{lags}}; \code{\link{last}};
  \code{\link{nperiod}}; \code{\link{null_frame}};
  \code{\link{null_join}}; \code{\link{rand}};
  \code{\link{relcumsum}}; \code{\link{relcum}};
  \code{\link{repl.inf.data.frame}}; \code{\link{repl.na}};
  \code{\link{safe_divide}}; \code{\link{send_email}};
  \code{\link{sort.abs}}; \code{\link{split.data}};
  \code{\link{unload}}; \code{\link{vsub}};
  \code{\link{which.inf}}; \code{\link{which.na}}
}

