% Generated by roxygen2 (4.0.2): do not edit by hand
\name{samp}
\alias{samp}
\title{Sample n rows from a dataframe. Works just like head/tail.}
\usage{
samp(x, n = 6L)
}
\arguments{
\item{x}{Dataframe to sample.}

\item{n}{Number of rows to sample.}
}
\value{
Dataframe with n randomly selected rows.
}
\description{
Sample n rows from a dataframe. Works just like head/tail.
}
\seealso{
Other utils: \code{\link{any.na}}; \code{\link{last}};
  \code{\link{percent}}; \code{\link{send_email}};
  \code{\link{which.na}}
}

