% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{samp}
\alias{samp}
\title{Sample n rows from a dataframe. Works just like head/tail.}
\usage{
samp(x, n = 6L)
}
\arguments{
\item{x}{Dataframe to sample.}

\item{n}{Number of rows to sample.}
}
\value{
Dataframe with n randomly selected rows.
}
\description{
Sample n rows from a dataframe. Works just like head/tail.
}
\seealso{
Other utils: \code{\link{any.na}};
  \code{\link{as.numeric.cols}}; \code{\link{chunk.ds}};
  \code{\link{coalesce}}; \code{\link{format.percent}};
  \code{\link{last}}; \code{\link{relcumsum}};
  \code{\link{relcum}}; \code{\link{send.email}};
  \code{\link{sort.abs}}; \code{\link{split.data}};
  \code{\link{which.na}}
}

