% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{sample.var}
\alias{sample.var}
\title{Sample variance (unbiased)}
\usage{
sample.var(x, x2)
}
\arguments{
\item{x}{data vector}

\item{x2}{data vector squared}
}
\description{
Sample variance (unbiased)
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{center}}; \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{rescale}}; \code{\link{sphere}};
  \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

