% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{send.email}
\alias{send.email}
\title{Wrapper for sending email using mutt.}
\usage{
send.email(to, subject, body, attachment)
}
\description{
Wrapper for sending email using mutt.
}
\seealso{
Other utils: \code{\link{any.na}}; \code{\link{chunk.ds}};
  \code{\link{format.percent}}; \code{\link{last}};
  \code{\link{samp}}; \code{\link{split.data}};
  \code{\link{which.na}}
}

