% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{send.email}
\alias{send.email}
\title{Wrapper for sending email using mutt.}
\usage{
send.email(to = NULL, subject = NULL, body = NULL, attachment = NULL)
}
\description{
Wrapper for sending email using mutt.
}
\seealso{
Other utils: \code{\link{any.na}};
  \code{\link{as.numeric.cols}}; \code{\link{chunk.ds}};
  \code{\link{coalesce}}; \code{\link{format.percent}};
  \code{\link{last}}; \code{\link{relcumsum}};
  \code{\link{relcum}}; \code{\link{samp}};
  \code{\link{sort.abs}}; \code{\link{split.data}};
  \code{\link{which.na}}
}

