% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{send_email}
\alias{send_email}
\title{Wrapper for sending email using mutt.}
\usage{
send_email(to = NULL, cc = NULL, subject = NULL, body = NULL,
  attachment = NULL)
}
\arguments{
\item{to}{}

\item{cc}{}

\item{subject}{}

\item{body}{}

\item{attachment}{}
}
\description{
Wrapper for sending email using mutt.
}

