% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{sort.abs}
\alias{sort.abs}
\title{Sort by absolute value using base::sort}
\usage{
\method{sort}{abs}(x, ...)
}
\arguments{
\item{x}{vector be sorted}
}
\description{
Sort by absolute value using base::sort
}
\seealso{
Other utils: \code{\link{any.na}};
  \code{\link{as.numeric.cols}}; \code{\link{chunk.ds}};
  \code{\link{coalesce}}; \code{\link{format.percent}};
  \code{\link{last}}; \code{\link{relcumsum}};
  \code{\link{relcum}}; \code{\link{samp}};
  \code{\link{send.email}}; \code{\link{split.data}};
  \code{\link{which.na}}
}

