% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{sphere}
\alias{sphere}
\title{Sphere data. This is a stronger enforcement than scaling as it forces Cov = I.}
\usage{
sphere(x)
}
\arguments{
\item{x}{data matrix to be sphered}
}
\description{
Sphere data. This is a stronger enforcement than scaling as it forces Cov = I.
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{center}}; \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{ma}}; \code{\link{rescale}};
  \code{\link{sample.var}};
  \code{\link{spline.derivative}};
  \code{\link{summarize.coefficient.matrix}}
}

