% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.r
\name{summarize.coefficient.matrix}
\alias{summarize.coefficient.matrix}
\title{Summarize a coefficient matrix}
\usage{
summarize.coefficient.matrix(beta)
}
\arguments{
\item{beta}{matrix of coefficient where rows represent variables and columns represent hyperparameter}
}
\description{
Summarize a coefficient matrix
}
\seealso{
Other stats: \code{\link{add.bias}}; \code{\link{bstrap}};
  \code{\link{center}}; \code{\link{emperical.derivative}};
  \code{\link{importance.cv.glmnet}};
  \code{\link{importance.glmnet}};
  \code{\link{importance.lm}}; \code{\link{importance}};
  \code{\link{ma}}; \code{\link{rescale}};
  \code{\link{sample.var}}; \code{\link{sphere}};
  \code{\link{spline.derivative}}
}

