% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{unload}
\alias{unload}
\title{unload}
\usage{
unload(package)
}
\description{
Unload a package using library syntax.
See: http://stackoverflow.com/questions/6979917/how-to-unload-a-package-without-restarting-r
}

