% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.r
\name{var_sub}
\alias{var_sub}
\title{substitute variables in string}
\usage{
var_sub(string, ...)
}
\arguments{
\item{string}{input string with variable}

\item{...}{variables to replaced of the form "blah" = 5}
}
\description{
substitute variables in string
}

