% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{which.na}
\alias{which.na}
\title{Subset a vector to its values which are NA.}
\usage{
which.na(x)
}
\arguments{
\item{x}{Vector with missing values.}
}
\value{
Subset of vector without missing values.
}
\description{
Subset a vector to its values which are NA.
}
\seealso{
Other utils: \code{\link{any.na}}; \code{\link{last}};
  \code{\link{percent}}; \code{\link{samp}};
  \code{\link{send.email}}; \code{\link{split.data}}
}

